+++ merlin/celery.py
@@ -6,7 +6,7 @@
 #
 # LLNL-CODE-797170
 # All rights reserved.
-# This file is part of Merlin, Version: 1.5.2.
+# This file is part of Merlin, Version: 1.6.1.
 #
 # For details, see https://github.com/LLNL/merlin.
 #
@@ -47,6 +47,7 @@ from merlin.config import (
     broker,
     results_backend,
 )
+from merlin.config.configfile import CONFIG
 from merlin.log_formatter import FORMATS
 from merlin.router import route_for_task
 
@@ -100,9 +101,12 @@ app.conf.update(
     redis_max_connections=100000,
 )
 
-# Set a one hour timeout to acknowledge a task before it's available to grab
-# again.
-app.conf.broker_transport_options = {"visibility_timeout": 7200, "max_connections": 100}
+# Set a timeout to acknowledge a task before it's available to grab
+# again (default 24 hours).
+app.conf.broker_transport_options = {
+    "visibility_timeout_seconds": CONFIG.celery.visibility_timeout_seconds,
+    "max_connections": 100,
+}
 
 app.conf.update(broker_pool_limit=0)